refactor: remove unused auto_subdir parameter from _process_add_configuration
The auto_subdir parameter in _process_add_configuration() was not used anywhere
in the function body - it was just a pass-through wrapper that didn't forward
the auto_subdir parameter to _prepare_add_configuration. Since this is a private
function, we can safely remove the unused parameter entirely.

Updated both:
- Function signature: removed auto_subdir parameter
- Call site: removed auto_subdir argument from function call

This results in a cleaner, more focused function interface.